made everything -wall and -Wextra compliant

+++ src/games/strategy_io.cpp
@@ -12,7 +12,7 @@ std::vector<double> StrategyIO::ReadIdStrategyIntegerIds(std::string &filename) 
   std::vector<double> strategy;
 
   std::ifstream infile(filename);
-  int id;
+  size_t id;
   double probability;
   std::string line;
   while (std::getline(infile, line)) {
@@ -20,7 +20,7 @@ std::vector<double> StrategyIO::ReadIdStrategyIntegerIds(std::string &filename) 
     std::vector<std::string> split_line = StringFunctions::SplitWithQuotes(line, ':');
     id = stoi(split_line[0]);
     probability = stof(split_line[1]);
-    if (id != strategy.size()) printf("StrategyReader::ReadIdStrategyIntegerIds error: id and vector size do not match. id: %d, size: %lu", id, strategy.size());
+    if (id != strategy.size()) printf("StrategyReader::ReadIdStrategyIntegerIds error: id and vector size do not match. id: %lu, size: %lu", id, strategy.size());
     strategy.push_back(std::max(probability, 0.0));
   }
   infile.close();
@@ -31,7 +31,7 @@ std::vector<double> StrategyIO::ReadStrategyStringIds(const std::string &filenam
                                                       const std::vector<std::string> &sequence_names,
                                                       int num_sequences) {
   std::map<std::string, int>  map;
-  for (int seq = 0; seq < sequence_names.size(); seq++) {
+  for (size_t seq = 0; seq < sequence_names.size(); seq++) {
     map[sequence_names[seq]] = seq;
   }
   std::vector<double> strategy((unsigned long) num_sequences);
